refactor: split test in two files
